[BUGFIX] Links of the tagcloud are broken
If you click on a link in the tagcloud of the master branch you will get an
TYPO3 exception like:

but was of type "NULL".

This is a result of a mistake in the tagcloud partial. The arguments in
the link action are wrong. This patch will solve this.

Change-Id: If3a1424242524c9d77f3849038356f52ef4d99b6
Resolves: #39151
Releases: 6.0