[BUGFIX] Backend module does not work
In the actual backend modul are some bugs. The new and edit action
templates uses the outdated f:form.textbox viewhelper. And because of
this you can save a new blog without entering the requiered title.

The markup of the f:form.select was outdated to and if you don't have
a administrator set you get an error after saving the blog.

This patch solves these promblems.

Fixes: #41742
Releases: 6.0

Change-Id: I18c74084bb706063304c7fa6703b00a5f9afb92e